native: move rejection shapes to their own module (line budget); the caller-reference 404 needs positive evidence (Greptile on #851)
