+++ exp/runtime/gateway/native/src/lib.rs
@@ -18,6 +18,7 @@ mod guardrails;
 mod memory;
 mod metrics;
 mod param_attribution;
+mod rejection_shapes;
 mod relay;
 mod replay;
 mod respond;
